rcx_unpack: ignore the info in the ASCII header
The body consists of concatenated EPSON IPL blocks.

In my original example the ASCII header had info on how long the blocks
are and I used that to split them, but not all firmwares have that info.

Now, instead, use the header information of each EPSON IPL block to know
its length, which tells us where the next one is.